@@ -28,85 +16,104 @@
     "stackdriver logging",
     "stackdriver"
   ],
+  "repository": "googleapis/nodejs-logging",
+  "license": "Apache-2.0",
+  "author": "Google Inc.",
+  "main": "build/src/index.js",
+  "files": [
+    "build/src",
+    "build/protos"
+  ],
   "scripts": {
-    "test": "c8 mocha --recursive ./build/test",
-    "docs": "jsdoc -c .jsdoc.js",
-    "predocs": "npm run compile",
-    "presystem-test": "npm run compile",
-    "system-test": "c8 mocha build/system-test --timeout 600000",
-    "samples-test": "cd samples/ && c8 npm test && cd ../",
-    "lint": "gts check && eslint '**/*.js'",
     "check": "gts check",
     "clean": "gts clean",
-    "compile": "tsc -p . && cp -r src/v2 build/src/v2 && cp -r protos build && cp test/*.js build/test",
-    "fix": "gts fix && eslint --fix '**/*.js'",
-    "prepare": "npm run compile",
-    "pretest": "npm run compile",
-    "posttest": "npm run check",
-    "docs-test": "linkinator docs",
+    "compile": "tsc -p . && cp -r proto* build/",
+    "compile-protos": "compileProtos src",
+    "predocs": "npm run compile",
+    "docs": "jsdoc -c .jsdoc.js",
     "predocs-test": "npm run docs",
-    "prelint": "cd samples; npm link ../; npm i"
+    "docs-test": "linkinator docs",
+    "fix": "gts fix && eslint '**/*.js' --fix",
+    "prelint": "cd samples; npm link ../; npm i",
+    "lint": "gts fix && eslint samples/*.js",
+    "prepare": "npm run compile",
+    "samples-test": "cd samples/ && npm link ../ && rm -rf node_modules && npm install && npm test && cd ../",
+    "presystem-test": "npm run compile",
+    "system-test": "mocha build/system-test --timeout 600000",
+    "pretest": "npm run compile",
+    "test": "c8 mocha build/test",
+    "posttest": "npm run check"
   },
   "dependencies": {
-    "@google-cloud/common": "^2.2.2",
-    "@google-cloud/paginator": "^2.0.0",
-    "@google-cloud/projectify": "^1.0.0",
-    "@google-cloud/promisify": "^1.0.0",
+    "@google-cloud/common": "^2.4.0",
+    "@google-cloud/paginator": "^2.0.3",
+    "@google-cloud/projectify": "^1.0.4",
+    "@google-cloud/promisify": "^1.0.4",
     "@opencensus/propagation-stackdriver": "0.0.20",
-    "arrify": "^2.0.0",
-    "dot-prop": "^5.1.0",
+    "arrify": "^2.0.1",
+    "dot-prop": "^5.2.0",
     "eventid": "^1.0.0",
     "extend": "^3.0.2",
-    "gcp-metadata": "^3.1.0",
-    "google-auth-library": "^5.2.2",
-    "google-gax": "^1.11.0",
+    "gcp-metadata": "^3.5.0",
+    "google-auth-library": "^5.10.1",
+    "google-gax": "^1.15.1",
     "is": "^3.3.0",
     "on-finished": "^2.3.0",
-    "pumpify": "^2.0.0",
-    "snakecase-keys": "^3.0.0",
-    "stream-events": "^1.0.4",
-    "through2": "^3.0.0",
+    "pumpify": "^2.0.1",
+    "snakecase-keys": "^3.1.2",
+    "stream-events": "^1.0.5",
+    "through2": "^3.0.1",
     "type-fest": "^0.12.0"
   },
   "devDependencies": {
-    "@google-cloud/bigquery": "^4.0.0",
-    "@google-cloud/pubsub": "^1.0.0",
-    "@google-cloud/storage": "^4.0.0",
-    "@types/extend": "^3.0.0",
+    "@google-cloud/bigquery": "^4.7.0",
+    "@google-cloud/pubsub": "^1.6.0",
+    "@google-cloud/storage": "^4.6.0",
+    "@types/extend": "^3.0.1",
     "@types/is": "0.0.21",
-    "@types/mocha": "^7.0.0",
+    "@types/mocha": "^7.0.2",
     "@types/mv": "^2.1.0",
-    "@types/ncp": "^2.0.1",
+    "@types/ncp": "^2.0.3",
+    "@types/node": "^13.9.2",
     "@types/on-finished": "^2.3.1",
     "@types/proxyquire": "^1.3.28",
     "@types/pumpify": "^1.4.1",
-    "@types/sinon": "^7.0.8",
+    "@types/sinon": "^7.5.2",
     "@types/through2": "^2.0.34",
     "@types/tmp": "^0.1.0",
-    "@types/uuid": "^7.0.0",
+    "@types/uuid": "^7.0.2",
     "assert-rejects": "^1.0.0",
     "bignumber.js": "^9.0.0",
-    "c8": "^7.0.0",
-    "codecov": "^3.0.4",
-    "eslint": "^6.0.0",
-    "eslint-config-prettier": "^6.0.0",
+    "c8": "^7.1.0",
+    "codecov": "^3.6.5",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^3.0.0",
+    "eslint-plugin-prettier": "^3.1.2",
     "execa": "^4.0.0",
-    "gts": "^1.0.0",
+    "gts": "^1.1.2",
     "http2spy": "^1.1.0",
-    "jsdoc": "^3.6.2",
-    "jsdoc-fresh": "^1.0.1",
-    "jsdoc-region-tag": "^1.0.2",
-    "linkinator": "^2.0.0",
-    "mocha": "^7.0.0",
+    "jsdoc": "^3.6.3",
+    "jsdoc-fresh": "^1.0.2",
+    "jsdoc-region-tag": "^1.0.4",
+    "linkinator": "^2.0.3",
+    "mocha": "^7.1.1",
     "mv": "^2.1.1",
     "ncp": "^2.0.0",
-    "nock": "^12.0.0",
-    "prettier": "^1.15.1",
-    "proxyquire": "^2.1.0",
-    "sinon": "^9.0.0",
-    "typescript": "3.6.4",
-    "uuid": "^7.0.2"
+    "nock": "^12.0.3",
+    "null-loader": "^3.0.0",
+    "pack-n-play": "^1.0.0-2",
+    "power-assert": "^1.6.1",
+    "prettier": "^1.19.1",
+    "proxyquire": "^2.1.3",
+    "sinon": "^9.0.1",
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.0",
+    "uuid": "^7.0.2",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11"
+  },
+  "engines": {
+    "node": ">=8.10.0"
   }
 }
