@@ -1,56 +1,51 @@
-// Copyright 2019 Google LLC
+// Copyright 2020 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//     https://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+//
+// ** This file is automatically generated by gapic-generator-typescript. **
+// ** https://github.com/googleapis/gapic-generator-typescript **
+// ** All changes to this file may be overwritten. **
 
-import execa = require('execa');
-import * as mv from 'mv';
-import {ncp} from 'ncp';
-import * as tmp from 'tmp';
-import {promisify} from 'util';
+import {packNTest} from 'pack-n-play';
+import {readFileSync} from 'fs';
+import {describe, it} from 'mocha';
 
-const keep = false;
-const mvp = (promisify(mv) as {}) as (...args: string[]) => Promise<void>;
-const ncpp = promisify(ncp);
-const stagingDir = tmp.dirSync({keep, unsafeCleanup: true});
-const stagingPath = stagingDir.name;
-const pkg = require('../../package.json');
-
-describe('📦 pack and install', () => {
-  /**
-   * Create a staging directory with temp fixtures used to test on a fresh
-   * application.
-   */
-  it('should be able to use the d.ts', async () => {
-    await execa('npm', ['pack', '--unsafe-perm']);
-    const tarball = `google-cloud-logging-${pkg.version}.tgz`;
-    await mvp(tarball, `${stagingPath}/logging.tgz`);
-    await ncpp('system-test/fixtures/sample', `${stagingPath}/`);
-    await execa('npm', ['install', '--unsafe-perm'], {
-      cwd: `${stagingPath}/`,
-      stdio: 'inherit',
-    });
-    await execa('node', ['--throw-deprecation', 'build/src/index.js'], {
-      cwd: `${stagingPath}/`,
-      stdio: 'inherit',
-    });
+describe('typescript consumer tests', () => {
+  it('should have correct type signature for typescript users', async function() {
+    this.timeout(300000);
+    const options = {
+      packageDir: process.cwd(), // path to your module.
+      sample: {
+        description: 'typescript based user can use the type definitions',
+        ts: readFileSync(
+          './system-test/fixtures/sample/src/index.ts'
+        ).toString(),
+      },
+    };
+    await packNTest(options); // will throw upon error.
   });
 
-  /**
-   * CLEAN UP - remove the staging directory when done.
-   */
-  after('cleanup staging', () => {
-    if (!keep) {
-      stagingDir.removeCallback();
-    }
+  it('should have correct type signature for javascript users', async function() {
+    this.timeout(300000);
+    const options = {
+      packageDir: process.cwd(), // path to your module.
+      sample: {
+        description: 'typescript based user can use the type definitions',
+        ts: readFileSync(
+          './system-test/fixtures/sample/src/index.js'
+        ).toString(),
+      },
+    };
+    await packNTest(options); // will throw upon error.
   });
 });
