@@ -11,13 +11,11 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+//
+// ** This file is automatically generated by gapic-generator-typescript. **
+// ** https://github.com/googleapis/gapic-generator-typescript **
+// ** All changes to this file may be overwritten. **
 
-'use strict';
-
-const ConfigServiceV2Client = require('./config_service_v2_client');
-const LoggingServiceV2Client = require('./logging_service_v2_client');
-const MetricsServiceV2Client = require('./metrics_service_v2_client');
-
-module.exports.ConfigServiceV2Client = ConfigServiceV2Client;
-module.exports.LoggingServiceV2Client = LoggingServiceV2Client;
-module.exports.MetricsServiceV2Client = MetricsServiceV2Client;
+export {ConfigServiceV2Client} from './config_service_v2_client';
+export {LoggingServiceV2Client} from './logging_service_v2_client';
+export {MetricsServiceV2Client} from './metrics_service_v2_client';
